requirements.txt
Django>=3.0,<4.0
Pillow>=10.0.0
python-decouple>=3.8
gunicorn>=20.1.0
psycopg2-binary>=2.9.0
whitenoise>=6.0.0
dj-database-url>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Developer Tools
django-debug-toolbar>=4.2.0
django-extensions>=3.2.0
django-silk>=5.0.0